+++ resources/views/components/icon/minus-square-o.blade.php
@@ -1,0 +1,1 @@
+<svg {{ $attributes }} xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20"><path d="M0,0V20H20V0ZM19.333,19.333H.667V.667H19.333Z"/><rect width="14" height="0.667" transform="translate(3 9.667)"/></svg>