+++ resources/views/components/icon/plus-square-o.blade.php
@@ -1,0 +1,1 @@
+<svg {{ $attributes }} xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20"><path d="M0,0V20H20V0ZM19.333,19.333H.667V.667H19.333Z"/><path d="M15.667,23h.667V16.333H23v-.667H16.333V9h-.667v6.667H9v.667h6.667Z" transform="translate(-6 -6)"/></svg>