+++ stubs/resources/views/pages/home.blade.php
@@ -1,7 +1,6 @@
-<x-gotime-app-layout layout="{{ config('naykel.template') }}">
+<x-gotime-app-layout layout="{{ config('naykel.template') }}" hasContainer class="py-5-3-2">
 
-    <div class="container py-5-3-2">
-        <h1>Home Page</h1>
-    </div>
+    <h1>{{ $title ?? null }}</h1>
+
 
 </x-gotime-app-layout>